Update to modern rollout typescript config.
The `reactotron-react-native` package needed to use @rollup/plugin-sucrase instead of @rollup/plugin-typescript because of the way react-native works. Kept getting all kind of errors becuase of the external expendency on react-native.
