+++ lib/eslint-plugin-reactotron/package.json
@@ -48,11 +48,12 @@
     "@babel/core": "^7.23.2",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-typescript": "^7.23.2",
+    "@rollup/plugin-typescript": "11.1.1",
     "@types/jest": "^29.5.12",
     "@types/node": "^18.19.14",
     "@typescript-eslint/eslint-plugin": "^6.7.5",
     "@typescript-eslint/parser": "^6.7.5",
-    "@typescript-eslint/utils": "^6.14.0",
+    "@typescript-eslint/utils": "^6.21.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^29.7.0",
     "eslint": "^8.56.0",
@@ -69,10 +70,8 @@
     "reactotron-core-client": "workspace:*",
     "rimraf": "5.0.5",
     "rollup": "^4.9.6",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-babel-minify": "^10.0.0",
     "rollup-plugin-filesize": "^10.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-terser": "^7.0.2",
     "testdouble": "^3.20.1",
     "ts-jest": "^29.1.2",
     "typescript": "^4.9.5"
