+++ lib/eslint-plugin-reactotron/rollup.config.cjs
@@ -1,7 +1,6 @@
-import resolve from "rollup-plugin-node-resolve"
-import babel from "rollup-plugin-babel"
 import filesize from "rollup-plugin-filesize"
-import minify from "rollup-plugin-babel-minify"
+import { terser } from "rollup-plugin-terser"
+import typescript from "@rollup/plugin-typescript"
 
 const pkg = require("./package.json")
 
@@ -17,15 +16,6 @@ export default {
       format: "esm",
     },
   ],
-  plugins: [
-    resolve({ extensions: [".ts"] }),
-    babel({ extensions: [".ts"], runtimeHelpers: true }),
-    process.env.NODE_ENV === "production"
-      ? minify({
-          comments: false,
-        })
-      : null,
-    filesize(),
-  ],
+  plugins: [typescript(), process.env.NODE_ENV === "production" ? terser() : null, filesize()],
   external: ["reactotron-core-client"],
 }
