+++ lib/reactotron-core-client/rollup.config.cjs
@@ -1,35 +1,28 @@
-import resolve from "rollup-plugin-node-resolve"
-import babel from "rollup-plugin-babel"
-import replace from "rollup-plugin-replace"
+import replace from "@rollup/plugin-replace"
 import filesize from "rollup-plugin-filesize"
-import minify from "rollup-plugin-babel-minify"
-
-const coreClientVersion = require("./package.json").version
+import { terser } from "rollup-plugin-terser"
+import typescript from "@rollup/plugin-typescript"
+const pkg = require("./package.json")
 
 /** @type {import('rollup').RollupOptions} */
 export default {
   input: "src/reactotron-core-client.ts",
   output: [
     {
-      file: "dist/index.js",
+      file: pkg.main,
       format: "cjs",
     },
     {
-      file: "dist/index.esm.js",
+      file: pkg.module,
       format: "esm",
     },
   ],
   plugins: [
-    resolve({ extensions: [".ts"] }),
+    typescript(),
     replace({
-      REACTOTRON_CORE_CLIENT_VERSION: coreClientVersion,
+      REACTOTRON_CORE_CLIENT_VERSION: pkg.version,
     }),
-    babel({ extensions: [".ts"], runtimeHelpers: true }),
-    process.env.NODE_ENV === "production"
-      ? minify({
-          comments: false,
-        })
-      : null,
+    process.env.NODE_ENV === "production" ? terser() : null,
     filesize(),
   ],
 }
