+++ lib/reactotron-core-contract/package.json
@@ -44,8 +44,11 @@
   "devDependencies": {
     "@babel/core": "^7.23.2",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.23.9",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-typescript": "^7.23.2",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-typescript": "11.1.1",
     "@types/jest": "^29.5.12",
     "@types/node": "^18.19.14",
     "@types/ws": "^8.5.10",
@@ -65,11 +68,8 @@
     "prettier": "^3.0.3",
     "rimraf": "5.0.5",
     "rollup": "^4.9.6",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-babel-minify": "^10.0.0",
     "rollup-plugin-filesize": "^10.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-terser": "^7.0.2",
     "ts-jest": "^29.1.2",
     "typescript": "^4.9.5",
     "ws": "^8.16.0"
