+++ lib/reactotron-core-ui/package.json
@@ -68,6 +68,7 @@
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-react": "7.22.15",
     "@babel/preset-typescript": "^7.23.2",
+    "@rollup/plugin-typescript": "11.1.1",
     "@storybook/addon-actions": "6.1.21",
     "@storybook/addon-knobs": "6.2.9",
     "@storybook/addon-links": "6.1.21",
@@ -103,11 +104,9 @@
     "react-tooltip": "4.5.1",
     "rimraf": "5.0.5",
     "rollup": "^4.9.6",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-babel-minify": "^10.0.0",
     "rollup-plugin-filesize": "^10.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-terser": "^7.0.2",
     "styled-components": "^6.1.8",
     "ts-jest": "^29.1.2",
     "typescript": "^4.9.5"
