+++ lib/reactotron-core-ui/rollup.config.cjs
@@ -1,30 +1,25 @@
-import resolve from "rollup-plugin-node-resolve"
-import babel from "rollup-plugin-babel"
 import filesize from "rollup-plugin-filesize"
-import minify from "rollup-plugin-babel-minify"
+import { terser } from "rollup-plugin-terser"
 import external from "rollup-plugin-peer-deps-external"
+import typescript from "@rollup/plugin-typescript"
+const pkg = require("./package.json")
 
 export default {
   input: "src/index.ts",
   output: [
     {
-      file: "dist/index.js",
+      file: pkg.main,
       format: "cjs",
     },
     {
-      file: "dist/index.esm.js",
+      file: pkg.module,
       format: "esm",
     },
   ],
   plugins: [
-    resolve({ extensions: [".ts", ".tsx"] }),
+    typescript(),
     external({ includeDependencies: true }),
-    babel({ extensions: [".ts", ".tsx"], runtimeHelpers: true }),
-    process.env.NODE_ENV === "production"
-      ? minify({
-          comments: false,
-        })
-      : null,
+    process.env.NODE_ENV === "production" ? terser() : null,
     filesize(),
   ],
 }
