+++ lib/reactotron-react-js/package.json
@@ -29,8 +29,8 @@
     "format": "prettier '*.{js,ts,tsx,json,css,yml}|**/*.{js,ts,tsx,json,css,yml}'",
     "format:check": "yarn format --check",
     "format:write": "yarn format --write",
-    "build": "npm-run-all clean tsc compile",
-    "build:dev": "npm-run-all clean tsc compile:dev",
+    "build": "yarn tsc",
+    "build:dev": "yarn tsc",
     "clean": "rimraf ./dist",
     "lint": "eslint src --ext .ts,.tsx",
     "compile": "NODE_ENV=production rollup -c --strictDeprecations --bundleConfigAsCjs",
@@ -49,6 +49,8 @@
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-typescript": "^7.23.2",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-typescript": "11.1.1",
     "@types/react": "18.2.55",
     "@typescript-eslint/eslint-plugin": "^6.7.5",
     "@typescript-eslint/parser": "^6.7.5",
@@ -66,11 +68,8 @@
     "prettier": "^3.0.3",
     "rimraf": "5.0.5",
     "rollup": "^4.9.6",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-babel-minify": "^10.0.0",
     "rollup-plugin-filesize": "^10.0.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-terser": "^7.0.2",
     "ts-jest": "^29.1.2",
     "typescript": "^4.9.5"
   },
