+++ lib/reactotron-react-js/rollup.config.cjs
@@ -1,35 +1,28 @@
-import resolve from "rollup-plugin-node-resolve"
-import babel from "rollup-plugin-babel"
-import replace from "rollup-plugin-replace"
+import replace from "@rollup/plugin-replace"
 import filesize from "rollup-plugin-filesize"
-import minify from "rollup-plugin-babel-minify"
-
-const reactotronReactJsVersion = require("./package.json").version
+import { terser } from "rollup-plugin-terser"
+import typescript from "@rollup/plugin-typescript"
+const pkg = require("./package.json")
 
 export default {
   input: "src/index.ts",
   output: [
     {
-      file: "dist/index.js",
+      file: pkg.main,
       format: "cjs",
       exports: "named",
     },
     {
-      file: "dist/index.esm.js",
+      file: pkg.module,
       format: "esm",
     },
   ],
   plugins: [
-    resolve({ extensions: [".ts"] }),
+    typescript(),
     replace({
-      REACTOTRON_REACT_JS_VERSION: reactotronReactJsVersion,
+      REACTOTRON_REACT_JS_VERSION: pkg.version,
     }),
-    babel({ extensions: [".ts"], runtimeHelpers: true }),
-    process.env.NODE_ENV === "production"
-      ? minify({
-          comments: false,
-        })
-      : null,
+    process.env.NODE_ENV === "production" ? terser() : null,
     filesize(),
   ],
   external: ["stacktrace-js", "reactotron-core-client"],
