+++ lib/reactotron-react-js/tsconfig.json
@@ -1,18 +1,21 @@
 {
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
-    "rootDir": ".",
     "allowJs": false,
     "declaration": true,
     "declarationDir": "dist/types",
+    "rootDir": ".",
     "emitDeclarationOnly": true,
+    "emitDecoratorMetadata": true,
     "allowSyntheticDefaultImports": true,
+    "experimentalDecorators": true,
     "module": "esnext",
     "moduleResolution": "node",
     "noImplicitAny": false,
     "noImplicitReturns": true,
     "noImplicitThis": true,
     "noUnusedLocals": true,
+    "sourceMap": true,
     "target": "es5",
     "lib": ["dom", "es2017"],
     "types": ["react", "jest"],
