+++ lib/reactotron-react-native/rollup.config.cjs
@@ -1,17 +1,19 @@
-import resolve from "rollup-plugin-node-resolve"
-import babel from "rollup-plugin-babel"
 import filesize from "rollup-plugin-filesize"
-import minify from "rollup-plugin-babel-minify"
+import { terser } from "rollup-plugin-terser"
+import sucrase from "@rollup/plugin-sucrase"
+import resolve from "@rollup/plugin-node-resolve"
+const pkg = require("./package.json")
 
 function getPlugins() {
   return [
-    resolve({ extensions: [".ts", ".tsx"] }),
-    babel({ extensions: [".ts", ".tsx"], runtimeHelpers: true }),
-    process.env.NODE_ENV === "production"
-      ? minify({
-          comments: false,
-        })
-      : null,
+    resolve({
+      extensions: [".js", ".jsx", ".ts", ".tsx"],
+    }),
+    sucrase({
+      exclude: /^(.+\/)?node_modules\/.+$/,
+      transforms: ["typescript", "imports"],
+    }),
+    process.env.NODE_ENV === "production" ? terser() : null,
     filesize(),
   ]
 }
@@ -30,11 +32,11 @@ export default [
     input: "src/reactotron-react-native.ts",
     output: [
       {
-        file: "dist/index.js",
+        file: pkg.main,
         format: "cjs",
       },
       {
-        file: "dist/index.esm.js",
+        file: pkg.module,
         format: "esm",
       },
     ],
