+++ scripts/template/package.json
@@ -48,6 +48,7 @@
     "@babel/core": "^7.23.2",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-typescript": "^7.23.2",
+    "@rollup/plugin-typescript": "11.1.1",
     "@types/jest": "^29.5.7",
     "@types/node": "^18.18.8",
     "@typescript-eslint/eslint-plugin": "^6.7.5",
@@ -69,10 +70,8 @@
     "reactotron-core-client": "workspace:*",
     "rimraf": "5.0.5",
     "rollup": "^4.9.6",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-babel-minify": "^10.0.0",
     "rollup-plugin-filesize": "^10.0.0",
-    "rollup-plugin-node-resolve": "^4.0.0",
+    "rollup-plugin-terser": "^7.0.2",
     "ts-jest": "^29.1.1",
     "typescript": "^4.9.5"
   },
